refactor(base-nodes): make Compositor GPU-only, drop CPU fallback and duplicate node
Per "no fallbacks, always use WebGPU": fold the GPU compositing path into the
existing CompositorNode (nodetool.image.Compositor) instead of shipping a
separate CompositorGPU node alongside the Sharp one.

- CompositorNode now composites through the shared WebGPULayerCompositor via
  Node.js Dawn; Sharp is used only as the codec layer (decode → RGBA in,
  encode PNG out). The Sharp `composite()` blend path is removed.
- Removed the separate CompositorGPUNode and its exports.
- Undecodable layers are still skipped (input validation at the boundary), not
  a compute fallback.
- The compositor regression suite now requires a real WebGPU device and skips
  when none is available, since there is no CPU path to assert against.
