+++ packages/base-nodes/src/index.ts
@@ -299,7 +299,6 @@ export {
   ImageToImageNode,
   ImageEditorNode,
   CompositorNode,
-  CompositorGPUNode,
   PainterNode,
   IMAGE_NODES
 } from "./nodes/image.js";
